Moved a couple more generic types
