+++ src/lambda/lambdaTreeDataProvider.ts
@@ -14,17 +14,18 @@ import { registerCommand } from '../shared/telemetry/telemetryUtils'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/awsTreeProvider'
 import { AWSCommandTreeNode } from '../shared/treeview/nodes/awsCommandTreeNode'
 import { AWSTreeNodeBase } from '../shared/treeview/nodes/awsTreeNodeBase'
+// tslint:disable-next-line:ordered-imports
+import { DefaultRegionNode } from '../shared/treeview/nodes/defaultRegionNode'
 import { ErrorNode } from '../shared/treeview/nodes/errorNode'
+import { RegionNode } from '../shared/treeview/nodes/regionNode'
+import { showErrorDetails } from '../shared/treeview/webviews/showErrorDetails'
 import { intersection, toMap, updateInPlace } from '../shared/utilities/collectionUtils'
 import { localize } from '../shared/utilities/vsCodeUtils'
 import { deleteCloudFormation } from './commands/deleteCloudFormation'
 import { deleteLambda } from './commands/deleteLambda'
 import { invokeLambda } from './commands/invokeLambda'
-import { showErrorDetails } from './commands/showErrorDetails'
 import { CloudFormationStackNode } from './explorer/cloudFormationNodes'
-import { DefaultRegionNode } from './explorer/defaultRegionNode'
 import { FunctionNodeBase } from './explorer/functionNode'
-import { RegionNode } from './explorer/regionNode'
 import { StandaloneFunctionNode } from './explorer/standaloneNodes'
 import { configureLocalLambda } from './local/configureLocalLambda'
